Extract chat state and fetch logic into useChat hook
